requirements.txt
torch_geometric
tqdm
wandb
pandas
scikit-learn
matplotlib